Add files via upload
This SwiftUI-based to-do list app allows users to efficiently manage their tasks and stay organized. The app features an intuitive user interface where users can add, edit, and delete tasks, as well as mark them as completed. Designed with simplicity and functionality in mind, this project is ideal for anyone looking to track their daily tasks or explore SwiftUI development. The code is structured to be easily understandable and is a great starting point for further enhancements or customization.